+++ src/taskpane/App.vue
@@ -6,7 +6,7 @@ import Home from './views/Home.vue'
 import ReadMode from './views/ReadMode.vue'
 import ComposeMode from './views/ComposeMode.vue'
 import Settings from './views/Settings.vue'
-import ContactKnowledgeBase from './views/ContactKnowledgeBase.vue'
+import ContactProfile from './views/ContactProfile.vue'
 import { currentView, go } from './router'
 import { isSignedIn } from './composables/useAuth'
 
@@ -19,7 +19,7 @@ const components = {
   read: ReadMode,
   compose: ComposeMode,
   settings: Settings,
-  'contact-kb': ContactKnowledgeBase,
+  'contact-profile': ContactProfile,
 } as const
 </script>